cs: fixed tableReducer issues from SonarCloud
